+++ include/Registro.h
@@ -5,21 +5,21 @@
 class Registro {
 private:
     std::string notas;
-    int cuentaId;
-    int asientoId;
-    int id;
-    float debe,haber;
+    int64_t cuentaId;
+    int64_t asientoId;
+    int64_t id;
+    int64_t debe,haber;
 public:
-    Registro(int id, int cuentaId, int asientoId, std::string nota,float debe, float haber);
-    int getCuentaId();
-    int getAsientoId();
+    Registro(int64_t id, int64_t cuentaId, int64_t asientoId, std::string nota,int64_t debe, int64_t haber);
+    int64_t getCuentaId();
+    int64_t getAsientoId();
     std::string getNotas();
-    int getId();
-    float getDebe();
-    float getHaber();
+    int64_t getId();
+    int64_t getDebe();
+    int64_t getHaber();
     void setDebe();
     void setHaber();
-    void setId(int id);
+    void setId(int64_t id);
 };
 
 #endif
