+++ include/RegistroDAO.h
@@ -2,7 +2,7 @@
 #define REGISTRODAO_H
 
 #include <vector>
-
+#include <cstdint>
 class Registro;
 class sqlite3;
 
@@ -10,8 +10,8 @@ class RegistroDAO {
 
 public:
     static void guardar(sqlite3* db, Registro* registros);
-    static void eliminar(sqlite3* db, int id);
-    static void actualizar(sqlite3* db,int id);
+    static void eliminar(sqlite3* db, int64_t id);
+    static void actualizar(sqlite3* db, int64_t id);
     static void cargarRegistros(sqlite3* db, std::vector<Registro*> &registros);
     static std::vector<Registro*> cargarTodas(sqlite3* db);
 };
